common/requirements.txt
twine
keyrings.google-artifactregistry-auth
keyring